+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.0.8
 - Improved debug
 - Visual alert for player about record on/off
+- Ignores Tag Team final choice
+- Tag char limit
 
 # 0.0.7
 - You can set the current GM (in case you have multiple GMs)